collector/gathertown-active-users/requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
dnspython==2.2.1
exceptiongroup==1.0.4
gmail-connector==0.5.8
h11==0.14.0
idna==3.4
outcome==1.2.0
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
selenium==4.6.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
undetected-chromedriver==3.1.6
urllib3==1.26.12
websockets==10.4
wsproto==1.2.0